Finished string extraction for i18n
Also a minor bug in preview for moderator has been fixed: The
_moderator_userbar.html template contained the date in the "d M Y" format,
which means that a string with the name of the month would be written.
When the locale was 'el' this would contain a string with greek characters
something that would leat to a 500-error when the wagtailuserbar would be
called. This could be solved either by changign the date foramt to Y-m-d
(iso format, which is what i chose) or by adding a u before the '' in line
11 and 15 of wagtailuserbar.py.