Coalesce animated gifs before resizing them. Fixes #264
In an animated gif, each frame is only the size it needs to be to cover the changed part of the image.
This is often much lower than the actual image size and the Wand image backend was using this incorrect size in its resizing calculations, causing all animated gifs to be resized incorrectly or not at all.

This commit calls ImageMagicks "MagickCoalesceImages" on the wand image. This resizes every frame to the full size of the image so the resizing calculations will work properly.