`resolve_model_string` now raises `LookupError`
Django 1.7 raises a `LookupError` when looking for a model that does not
exist. This brings the code in to line with Django. The `LookupError`
from Django is allowed to propagate in 1.7, and a `LookupError` is now
raised instead of a `NameError` in 1.6.

All code using `resolve_model_string` has been changed to catch the new
errors.