Added system check for ForeignKey cascade
Pages can be linked to other Pages, Images, etc through ForeignKeys. If
the object being referenced by a page gets deleted, the page itself will
be deleted as well. Also, the page will not be cleanly removed from the
tree (leaving orphans, out of date num_child, etc). It is very common to
accidentally forget to set the on_delete action to models.SET_NULL.

This pull request checks all ForeignKey fields on pages and makes sure
that they are not set to cascade. If they are set to cascade, the
developer will be warned of their mistake.

This only works for Django 1.7 users as it uses the system checks
framework.