Perform page deletion on the base Page class, not page.specific
This avoids a treebeard bug where calling SpecificPage.delete() would fail to delete child pages that are not instances of SpecificPage.

This partially reverts #737, and adds logic at the model level to ensure that any other calls to SpecificPage.delete() get diverted to Page.delete() instead. Re the rationale given in #737, reverting this is legitimate (and does not cause any meaningful regression) because:
1) Django always fires pre/post delete signals from both Page and SpecificPage regardless of which one is deleted;
2) Overriding delete() can never reliably work whatever we do; even if we consistently call SpecificPage.delete() within wagtail, we can't get treebeard to do that for child pages because treebeard has no knowledge of how to obtain a SpecificPage instance.