Check both `page` and `self` in templatetag contexts
Old code that does not properly extend `Page.get_context()` for context
generation will not get the new `PAGE_TEMPLATE_VAR` in their contexts,
which did not play nicely with the `wagtailuserbar`. Now, the template
context is checked for both `PAGE_TEMPLATE_VAR` and `self`, in that
order, for a Page.