Remove special case allowing all page types to be children of Page.
This was required when Page was absent from PAGE_MODEL_CLASSES (due to
it being 'abstract' / non-creatable) and thus never appeared in a
parent_page_types list. Now it appears in PAGE_MODEL_CLASSES, and is
a valid parent page type for page types that do not specify
parent_page_types (or explicitly include Page in the list).