Fix image / document search for non-superusers.
Since it's difficult for search() queries to traverse arbitrary relations, the
instances_user_has_any_permission_for methods in wagtailcore.permission_policies.collections
have been simplified to build the list of candidate collections in a separate query, and
filter the final queryset with a `collection__in` clause. This means that we just need to
index the 'collection' field of image / document in order for search to work again.