Make preview window work on IE11
IE doesn't provide standard onload event listeners on opened windows, so this code did user-agent sniffing
on the string 'MSIE' to bypass the onload animation. This failed on IE11, which drops 'MSIE' from the
user agent string. The code now checks for the presence of addEventListener instead.

As a bonus, it now falls back on IE's (non-standard?) attachEvent method, so the animation now works on IE
after all.