Revert #3116 and fix #3076 by changing href="#" to href="javascript:void(0)" instead - fixes #3208
The stopPropagation() call was too eager, as this blocked legitimate clicks on the links in the 'more' dropdown.

Yes, I know javascript:void(0) is evil, but it's less evil than href="#" in that it doesn't have side effects. And if the design calls for an element that has the same styling and interactions as an `<a>` element but just happens to trigger an in-page action rather than navigating to a new URL, I don't see why I should jump through hoops undoing the native style of an alternative element like `<button>` to achieve something that's perfectly simple to do the 'evil' way, and as far as I'm aware has no negative effects other than to cause HTML purists to roll their eyes. SO THERE.