Fix tests for indexing functions to disregard operations that are triggered by post_save signals
test_removes_object wrongly asserted that the 'add' method (not 'delete') was called; this should have failed, but didn't, because the mock was picking up the 'add' call that happened during object creation as a result of the post_save signal. Here we insert calls to .reset_mock() to stop the signals from interfering with the tests, and fix the broken assertion that was masked by this bug.