Refactor EditView.post into is_valid and not is_valid code paths
A bit of intermediate refactoring to make it clearer what the next step is doing...

is_saving is true in the code paths where the form is valid and unlocked, so rename this condition to is_valid. cancel_workflow is the one action that is allowed while is_valid is false, so split that into identical "cancel workflow while valid" and "cancel workflow while invalid" cases. This way, all code paths are cleanly partitioned into is_valid and (not is_valid) branches, which we can reorganise in the next step.