Check if new pages are site roots and clear the site_root_paths cache (#6507)
Previously, we were only doing this check for existing pages as there
was no way for a new page to be a site root.

But since internationalisation was implemented, new pages can be site
roots if they have a translation that is assigned in a `Site` record's
`root_page` field.

This commit removes a condition in `Page.save()` so `.is_site_root()` is
now called on new pages as well and the cache can get cleared if needed.