Apply semver naming for minor release url
The current naming is conflicting with how the version number parts are
called accoring to semantic versioning. Wagtail does not strictly follow
semantic versioning, but it still seems reasonable to adopt the same
terminology as the rest of the world.

See related discussion in #7336 (in particular [this comment by @gasman](https://github.com/wagtail/wagtail/issues/7336#issuecomment-954223018))