Fix updatemodulepaths for Python 3.7
Fixes #8970. Python 3.7 incorrectly opens the stream for writing in text mode (https://github.com/python/cpython/commit/be6dbfb43b89989ccc83fbc4c5234f50f44c47ad), so work around this by writing the unicode string back rather than encoding back to bytes.

This will probably fail if we're on Windows AND Python 3.7 AND encounter non-ASCII characters (because in that case it'll be open for writing as Windows-1252 encoding), but it's probably the best we can do without abandoning the fileinput library entirely.