Restore autocomplete behaviour of search in chooser modals
Fixes #7720. Unfortunately the different search backends have diverged massively in the autocomplete APIs they support:

* Postgres FTS (and probably the other database FTS backends) ignores the `partial_match` kwarg on `.search()` and `SearchField`, and only supports the `.autocomplete()` / `AutocompleteField` API
* The database fallback backend (and probably Elasticsearch) accepts `partial_match` on `.search()` (in fact I think it always does partial matches regardless of that setting - untested), but raises `NotImplementedError` on `.autocomplete()`

As a result, to perform autocompletion across all backends, we need to try both methods, with a try/except. (Also, user-defined models such as snippets will need either or both of `AutocompleteField` or `SearchField(partial_match=True)` in `search_fields` depending on the backend in use.)

Additionally, while tests have been added to cover the autocomplete behaviour, the test suite only ever runs under the fallback backend (with the exception of the backend-specific tests in wagtail.search), and so these would have succeeded anyhow. Much more cleanup work is needed here, but this will serve as a stopgap solution to get the choosers working as intended again.